requirements.txt
# backend
fastapi
uvicorn[standard]
python-dotenv
python-multipart
requests
pandas
docker